requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.8.1
    # via django
certifi==2024.8.30
    # via requests
charset-normalizer==3.4.0
    # via requests
django==5.1.2
    # via
    #   -r requirements.in
    #   django-filter
    #   django-rest-knox
    #   djangorestframework
django-environ==0.11.2
    # via -r requirements.in
django-filter==24.3
    # via -r requirements.in
django-rest-knox==5.0.2
    # via -r requirements.in
djangorestframework==3.15.2
    # via
    #   -r requirements.in
    #   django-rest-knox
djangorestframework-camel-case==1.4.2
    # via -r requirements.in
gunicorn==21.2.0
    # via -r requirements.in
idna==3.10
    # via requests
packaging==24.2
    # via gunicorn
psycopg2==2.9.10
    # via -r requirements.in
requests==2.32.3
    # via -r requirements.in
sqlparse==0.5.1
    # via django
tzdata==2024.2
    # via django
urllib3==2.2.3
    # via requests
whitenoise==6.8.2
    # via -r requirements.in
